run decryption test on its own and last
